refactor(search): migrate to new MCP server architecture with static metadata
Migrate the search MCP server to use static metadata while preserving
its dynamic tool registration behavior (tags vs no-tags mode).

- Create metadata.ts with createToolsRecord for type-safe tool definitions
- Create tools/index.ts with searchFunction handler
- Create index.ts preserving dynamic tool registration logic
- Update constants.ts to use metadata field
- Update servers/index.ts to import from new location
- Move FIND_TAGS_TOOL_NAME to find_tags/metadata.ts to avoid circular deps
