+++ front/lib/actions/mcp_internal_actions/servers/data_sources_file_system.ts
@@ -5,7 +5,6 @@ import {
   FILESYSTEM_FIND_TOOL_NAME,
   FILESYSTEM_LIST_TOOL_NAME,
   FILESYSTEM_LOCATE_IN_TREE_TOOL_NAME,
-  FIND_TAGS_TOOL_NAME,
   SEARCH_TOOL_NAME,
 } from "@app/lib/actions/mcp_internal_actions/constants";
 import { registerCatTool } from "@app/lib/actions/mcp_internal_actions/tools/data_sources_file_system/cat";
@@ -17,6 +16,7 @@ import { shouldAutoGenerateTags } from "@app/lib/actions/mcp_internal_actions/to
 import { makeInternalMCPServer } from "@app/lib/actions/mcp_internal_actions/utils";
 import type { AgentLoopContextType } from "@app/lib/actions/types";
 import { registerFindTagsTool } from "@app/lib/api/actions/tools/find_tags";
+import { FIND_TAGS_TOOL_NAME } from "@app/lib/api/actions/tools/find_tags/metadata";
 import type { Authenticator } from "@app/lib/auth";
 
 function createServer(
