+++ front/lib/api/actions/tools/find_tags/index.ts
@@ -3,13 +3,13 @@ import type { TextContent } from "@modelcontextprotocol/sdk/types.js";
 import trim from "lodash/trim";
 
 import { MCPError } from "@app/lib/actions/mcp_errors";
-import { FIND_TAGS_TOOL_NAME } from "@app/lib/actions/mcp_internal_actions/constants";
 import type { DataSourcesToolConfigurationType } from "@app/lib/actions/mcp_internal_actions/input_schemas";
 import { getCoreSearchArgs } from "@app/lib/actions/mcp_internal_actions/tools/utils";
 import { withToolLogging } from "@app/lib/actions/mcp_internal_actions/wrappers";
 import type { AgentLoopContextType } from "@app/lib/actions/types";
 import {
   FIND_TAGS_BASE_DESCRIPTION,
+  FIND_TAGS_TOOL_NAME,
   findTagsSchema,
 } from "@app/lib/api/actions/tools/find_tags/metadata";
 import config from "@app/lib/api/config";
